tests: make sure some essential tests are run
